Simplify persistence helpers and keep texture transactions together
