+++ src/MaterialEditor.Base/MaterialEditor.Base.projitems
@@ -11,8 +11,6 @@
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)UI\Sections\Properties\PropertyRowBuilders.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialTextureImportTransaction.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialTextureSnapshot.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialTextureImportCommit.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialCubemapImportTransaction.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialEditResult.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\MaterialEditRequestQueue.cs" />
