+++ src/MaterialEditor.Base/Operations/MaterialTextureImportTransaction.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 
 namespace MaterialEditorAPI
@@ -38,6 +39,69 @@ namespace MaterialEditorAPI
                     try { discard(candidate); }
                     catch (Exception ex) { MaterialEditorPluginBase.Logger?.LogWarning("Texture cleanup: " + ex.Message); }
                 }
+            }
+        }
+
+        // Keep the existing DTO identity: animation bindings may still reference it.
+        internal static void Commit<T, TController, TAnimation>(IList<T> records, IDictionary<T, TController> animations,
+            T existing, T candidate, Func<T, int?> getId, Action<T, int?> setId,
+            Func<T, TAnimation> getAnimation, Action<T, TAnimation> setAnimation) where T : class
+        {
+            if (existing == null)
+            {
+                records.Add(candidate);
+                return;
+            }
+            var previousId = getId(existing);
+            var previousAnimation = getAnimation(existing);
+            TController previousController;
+            TController candidateController;
+            var hadPrevious = animations.TryGetValue(existing, out previousController);
+            var hasCandidate = animations.TryGetValue(candidate, out candidateController);
+            try
+            {
+                setId(existing, getId(candidate));
+                setAnimation(existing, getAnimation(candidate));
+                animations.Remove(candidate);
+                if (hasCandidate) animations[existing] = candidateController;
+                else animations.Remove(existing);
+            }
+            catch
+            {
+                setId(existing, previousId);
+                setAnimation(existing, previousAnimation);
+                animations.Remove(candidate);
+                if (hadPrevious) animations[existing] = previousController;
+                else animations.Remove(existing);
+                throw;
+            }
+        }
+    }
+
+    /// <summary>Rollback values for every runtime material touched by SetTexture, including projectors.</summary>
+    internal sealed class MaterialTextureSnapshot
+    {
+        private readonly Dictionary<Material, Texture> _values = new Dictionary<Material, Texture>();
+        private readonly string _property;
+
+        internal MaterialTextureSnapshot(GameObject root, string name, string property)
+        {
+            _property = "_" + property;
+            foreach (var material in MaterialAPI.GetObjectMaterials(root, name))
+                if (material != null && material.HasProperty(_property))
+                    _values[material] = material.GetTexture(_property);
+        }
+
+        internal void Restore()
+        {
+            foreach (var entry in _values)
+            {
+                try
+                {
+                    if (entry.Key != null && entry.Key.HasProperty(_property))
+                        entry.Key.SetTexture(_property, entry.Value);
+                }
+                catch (Exception ex) { MaterialEditorPluginBase.Logger?.LogWarning("Texture rollback: " + ex.Message); }
             }
         }
     }
