+++ src/MaterialEditor.Core/Core.MaterialEditor.CharaController.Persistence.cs
@@ -73,27 +73,27 @@ namespace KK_Plugins.MaterialEditor
                 else
                     data.data.Add(TexDicSaveKey, null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(RendererPropertyList), coordinateRendererPropertyList);
+                data.data.Add(nameof(RendererPropertyList), coordinateRendererPropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateRendererPropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(ProjectorPropertyList), coordinateProjectorPropertyList);
+                data.data.Add(nameof(ProjectorPropertyList), coordinateProjectorPropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateProjectorPropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialNamePropertyList), coordinateMaterialNamePropertyList);
+                data.data.Add(nameof(MaterialNamePropertyList), coordinateMaterialNamePropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialNamePropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialFloatPropertyList), coordinateMaterialFloatPropertyList);
+                data.data.Add(nameof(MaterialFloatPropertyList), coordinateMaterialFloatPropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialFloatPropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialKeywordPropertyList), coordinateMaterialKeywordPropertyList);
+                data.data.Add(nameof(MaterialKeywordPropertyList), coordinateMaterialKeywordPropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialKeywordPropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialColorPropertyList), coordinateMaterialColorPropertyList);
+                data.data.Add(nameof(MaterialColorPropertyList), coordinateMaterialColorPropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialColorPropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialVectorPropertyList), coordinateMaterialVectorPropertyList);
+                data.data.Add(nameof(MaterialVectorPropertyList), coordinateMaterialVectorPropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialVectorPropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialTexturePropertyList), coordinateMaterialTexturePropertyList);
+                data.data.Add(nameof(MaterialTexturePropertyList), coordinateMaterialTexturePropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialTexturePropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialCubemapPropertyList), coordinateMaterialCubemapPropertyList);
+                data.data.Add(nameof(MaterialCubemapPropertyList), coordinateMaterialCubemapPropertyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialCubemapPropertyList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialShaderList), coordinateMaterialShaderList);
+                data.data.Add(nameof(MaterialShaderList), coordinateMaterialShaderList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialShaderList) : null);
 
-                MaterialEditRecordSerialization.Write(data, nameof(MaterialCopyList), coordinateMaterialCopyList);
+                data.data.Add(nameof(MaterialCopyList), coordinateMaterialCopyList.Count > 0 ? MessagePackSerializer.Serialize(coordinateMaterialCopyList) : null);
 
                 SetCoordinateExtendedData(coordinate, data);
             }
@@ -432,8 +432,124 @@ namespace KK_Plugins.MaterialEditor
         /// </summary>
         private int SetAndGetTextureID(byte[] textureBytes)
         {
-            return MaterialTextureDataStore.GetOrAdd(TextureDictionary, textureBytes);
+            return TextureSaveHandler.GetOrAddTexture(TextureDictionary, textureBytes);
         }
 
+        private List<ObjectType> PrepareCharacterLoad()
+        {
+            List<ObjectType> objectTypesToLoad = new List<ObjectType>();
+
+            var loadFlags = MakerAPI.GetCharacterLoadFlags();
+            if (loadFlags == null)
+            {
+                RendererPropertyList.Clear();
+                ProjectorPropertyList.Clear();
+                MaterialNamePropertyList.Clear();
+                MaterialFloatPropertyList.Clear();
+                MaterialKeywordPropertyList.Clear();
+                MaterialColorPropertyList.Clear();
+                MaterialVectorPropertyList.Clear();
+                MaterialTexturePropertyList.Clear();
+                MaterialCubemapPropertyList.Clear();
+                MaterialShaderList.Clear();
+                MaterialCopyList.Clear();
+                AnimationControllerMap.Clear();
+
+                objectTypesToLoad.Add(ObjectType.Accessory);
+                objectTypesToLoad.Add(ObjectType.Character);
+                objectTypesToLoad.Add(ObjectType.Clothing);
+                objectTypesToLoad.Add(ObjectType.Hair);
+            }
+            else
+            {
+                bool changed = false;
+
+                if (loadFlags.Face || loadFlags.Body)
+                {
+                    RemovePropertiesForLoad((type, coordinate) => type == ObjectType.Character, true);
+
+                    objectTypesToLoad.Add(ObjectType.Character);
+
+                    changed = true;
+                }
+                if (loadFlags.Clothes)
+                {
+                    RemovePropertiesForLoad((type, coordinate) => type == ObjectType.Clothing, true);
+                    objectTypesToLoad.Add(ObjectType.Clothing);
+
+                    RemovePropertiesForLoad((type, coordinate) => type == ObjectType.Accessory, true);
+                    objectTypesToLoad.Add(ObjectType.Accessory);
+
+                    changed = true;
+                }
+                if (loadFlags.Hair)
+                {
+                    RemovePropertiesForLoad((type, coordinate) => type == ObjectType.Hair, true);
+                    objectTypesToLoad.Add(ObjectType.Hair);
+
+                    changed = true;
+                }
+
+                if (changed)
+                {
+                    PurgeUnusedAnimation();
+                }
+            }
+
+            return objectTypesToLoad;
+        }
+
+        private List<ObjectType> PrepareCoordinateLoad()
+        {
+            List<ObjectType> objectTypesToLoad = new List<ObjectType>();
+
+            var loadFlags = MakerAPI.GetCoordinateLoadFlags();
+            if (loadFlags == null)
+            {
+                RemovePropertiesForLoad((type, coordinate) => (type == ObjectType.Clothing || type == ObjectType.Accessory) && coordinate == CurrentCoordinateIndex, false);
+
+                objectTypesToLoad.Add(ObjectType.Accessory);
+                objectTypesToLoad.Add(ObjectType.Clothing);
+
+                PurgeUnusedAnimation();
+            }
+            else
+            {
+                if (loadFlags.Clothes)
+                {
+                    RemovePropertiesForLoad((type, coordinate) => type == ObjectType.Clothing && coordinate == CurrentCoordinateIndex, false);
+                    objectTypesToLoad.Add(ObjectType.Clothing);
+                }
+                if (loadFlags.Accessories)
+                {
+                    RemovePropertiesForLoad((type, coordinate) => type == ObjectType.Accessory && coordinate == CurrentCoordinateIndex, false);
+                    objectTypesToLoad.Add(ObjectType.Accessory);
+                }
+
+                if (loadFlags.Clothes || loadFlags.Accessories)
+                {
+                    PurgeUnusedAnimation();
+                }
+            }
+
+            return objectTypesToLoad;
+        }
+
+        private void RemovePropertiesForLoad(Func<ObjectType, int, bool> remove, bool includeProjectors)
+        {
+            RendererPropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialNamePropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialFloatPropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialKeywordPropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialColorPropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialVectorPropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialTexturePropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialCubemapPropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialShaderList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            MaterialCopyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+            // Coordinate loading historically excludes projector records.
+            if (includeProjectors)
+                ProjectorPropertyList.RemoveAll(x => remove(x.ObjectType, x.CoordinateIndex));
+        }
     }
 }
