+++ src/MaterialEditor.Core/Core.MaterialEditor.CharaController.Restore.Properties.cs
@@ -181,5 +181,35 @@ namespace KK_Plugins.MaterialEditor
                 MaterialAPI.SetProjectorProperty(FindGameObject(property.ObjectType, property.Slot), property.ProjectorName, property.Property, float.Parse(property.Value));
             }
         }
+
+        private void RestoreMaterialTexturePropertyList(CharacterRestoreScope scope)
+        {
+            for (var i = 0; i < MaterialTexturePropertyList.Count; i++)
+            {
+                var property = MaterialTexturePropertyList[i];
+                if (!scope.Includes(property.ObjectType, property.CoordinateIndex, CurrentCoordinateIndex)) continue;
+                if (property.ObjectType == ObjectType.Character && !scope.Body) continue;
+                var go = FindGameObject(property.ObjectType, property.Slot);
+                if (Instance.CheckBlacklist(property.MaterialName, property.Property)) continue;
+
+                SetTextureWithProperty(go, property);
+                SetTextureOffset(go, property.MaterialName, property.Property, property.Offset);
+                SetTextureScale(go, property.MaterialName, property.Property, property.Scale);
+            }
+        }
+
+        private void RestoreMaterialCubemapPropertyList(CharacterRestoreScope scope)
+        {
+            for (var i = 0; i < MaterialCubemapPropertyList.Count; i++)
+            {
+                var property = MaterialCubemapPropertyList[i];
+                if (!scope.Includes(property.ObjectType, property.CoordinateIndex, CurrentCoordinateIndex)) continue;
+                if (property.ObjectType == ObjectType.Character && !scope.Body) continue;
+                var go = FindGameObject(property.ObjectType, property.Slot);
+                if (Instance.CheckBlacklist(property.MaterialName, property.Property)) continue;
+
+                SetCubemapWithProperty(go, property);
+            }
+        }
     }
 }
