+++ src/MaterialEditor.Core/Core.MaterialEditor.projitems
@@ -9,12 +9,8 @@
     <Import_RootNamespace>Core.MaterialEditor</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Persistence\MaterialEditRecordSerialization.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Persistence\MaterialTextureDataStore.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.CharaController.Restore.Properties.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.CharaController.Restore.Textures.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Persistence\MaterialEditLoadContext.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.CharaController.Persistence.Selection.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.CharaController.Persistence.Properties.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)AssemblyInfo.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core.MaterialEditor.CharaRepository.cs" />
